osx goodies
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v1_7/bzflag@1540 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/menus.cxx
@@ -3582,7 +3582,7 @@ void			ServerMenu::addToList(ServerItem& info, bool doCache)
   // update display
   char buffer [80];
   std::vector<std::string> args;
-  sprintf(buffer, "%d", servers.size());
+  sprintf(buffer, "%d", (int)servers.size());
   args.push_back(buffer);
   setStatus("Servers found: {1}", &args);
